+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `forms` and `title` to `FormSubmissionApprovalsResponse`
+- `title` to `FormApprovalFlowInstanceHistory` and `FormSubmissionApprovalResponse`
+
 ## [11.1.1] - 2026-07-28
 
 ### Dependencies